fix(split-pr338-02): restore type and complexity gates
